feat: moved getting config browser to testcafe-configuration.ts
