+++ src/utils/get-options/index.ts
@@ -5,7 +5,6 @@ import getVideoOptions from './video';
 import getMetaOptions from './meta';
 import getGrepOptions from './grep';
 import getCompilerOptions from './compiler';
-import getBrowsersOptions from './browser';
 
 export {
     getSSLOptions,
@@ -14,6 +13,5 @@ export {
     getVideoOptions,
     getMetaOptions,
     getGrepOptions,
-    getCompilerOptions,
-    getBrowsersOptions,
+    getCompilerOptions
 };
